Rename this package from acmart to acmart-fare
Rename away from scribble/acmart to scribble/acmart-fare, since the
former now names a package part of the official Racket distribution.
I encourage users to use the official package and contribute fixes to it.
This is available for backward compatibility with documents that have
not been migrated (yet, if ever).

Upgrade the style file from 1.28 to 1.48. Add support for fontsize option.

Fix the nest macro and update its users: () is not a valid context, (begin) is.

Update version from 0.2017.0 to 0.2017.1.